refactor(v1): fold ids.py into types (EnvId) + utils (env helpers)
ids.py was a thin v1 wrapper over verifiers.utils.install_utils. Split it into the v1
submodule: the EnvId type + its validator move to types.py (next to the other shared
types), and env_name/env_module/ensure_installed move to utils.py (small shared helpers,
beside the install_utils calls they wrap). EnvId stays a vf.* export; the helper functions
are no longer top-level vf exports (import from verifiers.v1.utils).

Breaking: vf.env_name / vf.ensure_installed are dropped from the top-level export
(import from verifiers.v1.utils instead). EnvId is unchanged.
