+++ verifiers/v1/harness.py
@@ -19,7 +19,7 @@ from pydantic_config import BaseConfig
 from verifiers.v1.clients import RolloutContext
 from verifiers.v1.decorators import discover_decorated, invoke
 from verifiers.v1.errors import ProgramError
-from verifiers.v1.ids import EnvId, env_name
+from verifiers.v1.utils import env_name
 from verifiers.v1.runtimes import (
     ProgramResult,
     Runtime,
@@ -28,7 +28,7 @@ from verifiers.v1.runtimes import (
 )
 from verifiers.v1.task import Task
 from verifiers.v1.trace import Trace
-from verifiers.v1.types import Messages
+from verifiers.v1.types import EnvId, Messages
 
 logger = logging.getLogger(__name__)
