+++ verifiers/v1/types.py
@@ -8,7 +8,7 @@ them into these models explicitly.
 
 from typing import Annotated, Any, Literal
 
-from pydantic import AliasChoices, BaseModel, ConfigDict, Field
+from pydantic import AfterValidator, AliasChoices, BaseModel, ConfigDict, Field
 from renderers.base import MultiModalData
 from typing_extensions import TypedDict
 
@@ -218,3 +218,22 @@ class SamplingConfig(BaseModel):
     max_tokens: int | None = Field(
         None, validation_alias=AliasChoices("max_tokens", "max_completion_tokens")
     )
+
+
+# --- ids ----------------------------------------------------------------------
+
+
+def _validate_env_id(env_id: str) -> str:
+    """Validate the id's shape — a hub id must be a well-formed ``org/name[@version]``; a
+    local id is any module name. Returns it unchanged (the value stays a plain ``str``)."""
+    from verifiers.utils.install_utils import is_hub_env, parse_env_id
+
+    if is_hub_env(env_id):
+        parse_env_id(env_id)  # raises ValueError on a malformed org/name[@version]
+    return env_id
+
+
+EnvId = Annotated[str, AfterValidator(_validate_env_id)]
+"""A taskset / harness / environment id — ``name``, ``org/name``, or ``org/name@version``. A
+plain validated ``str``; derive its package/module name with `env_name` / `env_module`
+(`verifiers.v1.utils`)."""
